requirements.txt
boto==2.38.0
Django==1.8.5
django-bakery==0.7.5
Naked==0.1.31
psycopg2==2.6.1
PyYAML==3.11
requests==2.8.0
six==1.10.0
wheel==0.24.0
